Rename edits.py to caliban.py
