+++ browser/blueprints.py
@@ -24,7 +24,7 @@ from werkzeug.exceptions import HTTPException
 from helpers import is_trk_file, is_npz_file
 from files import CalibanFile
 from models import Project
-from edits import TrackEdit, ZStackEdit
+from caliban import TrackEdit, ZStackEdit
 
 
 bp = Blueprint('caliban', __name__)  # pylint: disable=C0103
@@ -72,7 +72,7 @@ def upload_file(project_id):
     state = load_project_state(project)
     filename = state.file.filename
 
-    # Call function in edits.py to save data file and send to S3 bucket
+    # Call function in caliban.py to save data file and send to S3 bucket
     if is_trk_file(filename):
         state.action_save_track()
     elif is_npz_file(filename):
